requirements.txt
asgiref==3.11.1
async-timeout==5.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
crispy-bootstrap5==2026.3
cryptography==46.0.5
Django==5.0.6
django-allauth==65.14.3
django-ckeditor==6.7.3
django-ckeditor-5==0.2.20
django-cleanup==9.0.0
django-crispy-forms==2.5
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-nested-admin==4.1.6
django-redis==6.0.0
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-monkey-business==1.1.0
redis==7.4.0
requests==2.32.5
sib-api-v3-sdk==7.6.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
